+++ include/diamond_engine/config/Collider2DConfig.h
@@ -10,13 +10,13 @@ namespace diamond_engine {
 	public:
 		virtual ~Collider2DConfig() override = default;
 
-		GLuint getWidth() const;
+		GLfloat getWidth() const;
 
-		void setWidth(GLuint width);
+		void setWidth(GLfloat width);
 
-		GLuint getHeight() const;
+		GLfloat getHeight() const;
 
-		void setHeight(GLuint height);
+		void setHeight(GLfloat height);
 
 		ColliderType getColliderType() const;
 
@@ -28,7 +28,7 @@ namespace diamond_engine {
 
 	private:
 		ColliderType m_colliderType{ ColliderType::OBSTACLE };
-		GLuint m_width{ 0 };
-		GLuint m_height{ 0 };
+		GLfloat m_width{ 0 };
+		GLfloat m_height{ 0 };
 	};
 }
